scrapper_script/requirements.txt
Flask>=2.0.0
playwright>=1.30.0
phonenumbers>=8.12.0
supabase>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0 